Raspberry Pi 4/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
idna==3.10
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2